Add files via upload
The assignments given have an outflow of the following - Assignment 1- describing the statistical outputs of the dataset.
Assignment 2 - Calculating confidence intervals and doing a step wise analysis of validating the hypothesis made.
Assignment 3 - Scatter plots and regression analysis.
Assignment 4 - Creating residual plots and analyzing using the F and T tests. Creating histograms and bar plots to analzye the variation in the data set.
Assignment 5 - Grouping of data and analyzing many samples. Using ANOVA and Tu-key, Bonferroni and other techniques for adjusting the p value for better analysis.
Assignment 6 - Proportional analysis of the data given with respect to simple, multiple linear regression as well as ANOVA, ANCOVA etc..